requirements.txt
# Versions chosen so pip can use wheels on Windows/Python 3.12+ (no local compiler needed).
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
sqlalchemy>=2.0.40,<3
pypdf==5.1.0
python-docx==1.1.2
Pillow>=11.1.0,<12
ImageHash==4.3.1
scikit-learn>=1.6.1,<1.8
numpy>=2.1.0,<3
joblib>=1.4.2
pydantic-settings==2.6.1
